// Client/src/app/models/patient.ts
export class Patient {
    id?: string;
    name: string;
    fileNo: number;
    citizenId?: string;
    birthdate: Date;
    gender: number;
    nationality?: string;
    phoneNumber?: string;
    email: string;
    country?: string;
    city?: string;
    street?: string;
    address1?: string;
    address2?: string;
    contactPerson?: string;
    contactRelation?: string;
    contactPhone?: string;
    photoUrl: string;
    firstVisitDate: Date;
}